requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg2-binary==2.9.10
sqlparse==0.5.2
tzdata==2024.2
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
environs[django]==9.5.0
pillow==9.0.1
django-debug-toolbar